Update train.py and predict.py
Disabled the showing of each image, as the increasing number of images made it unfeasable. Added successful prediction pictures to show what the face recognition program is able to do when it runs correctly. However, there are many times where it cannot detect a face at all, even when one is clearly present.